Patch Browser updates
Fixed the patch loading issue for SR-JV80-19 by adding the drumkits of all the expansions to the total number of patches. More trivially, also added another row to the browser table to make it possible to access the last entry in the occassional situation where an expansion has 255 or more patches and 4 or more drumkits combined.